# deploy/postman/noema-openapi.yaml
openapi: "3.0.3"
info:
  title: Noema API
  description: Production-grade AI reasoning engine
  version: "1.0.0"
  contact:
    name: Noema
servers:
  - url: http://localhost:8000
    description: Local development
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Master API key. Empty = auth disabled (dev only).
  schemas:
...
    Requirement:
      type: object
      required: [category, description]
      properties:
        category:
          type: string
          maxLength: 50
        description:
          type: string
          maxLength: 2000
        priority:
          type: integer
          minimum: 1
          maximum: 10
          default: 5
        constraints:
          type: array
          items:
            type: string
          maxItems: 20
    TaskRequest:
      type: object
      required: [title, description]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          minLength: 1
          maxLength: 10000
        requirements:
          type: array
          items:
            $ref: "#/components/schemas/Requirement"
          maxItems: 50
        preferred_stack:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          nullable: true
        complexity:
          type: string
          enum: [simple, moderate, complex, extreme]
          default: moderate
        tags:
          type: array
          items:
            type: string
          maxItems: 20
        context:
          type: object
        task_id:
          type: string
          nullable: true
          description: Optional task ID for cancellation tracking
...
    SolutionResponse:
      type: object
      properties:
        id:
          type: string
        task_id:
          type: string
        title:
          type: string
        summary:
          type: string
        quality:
          type: string
        confidence:
          type: number
        stack_summary:
          type: string
        code_blocks_count:
          type: integer
        performance_notes:
          type: array
          items:
            type: string
        security_notes:
          type: array
          items:
            type: string
        thought_steps:
          type: integer
        duration_ms:
          type: number
    SolutionDetailResponse:
      type: object
      properties:
        id:
          type: string
        task_id:
          type: string
        title:
          type: string
        summary:
          type: string
        quality:
          type: string
        confidence:
          type: number
        architecture:
          type: object
          nullable: true
        stack:
          type: object
        code_blocks:
          type: array
          items:
            type: object
        deployment:
          type: object
        performance_notes:
          type: array
          items:
            type: string
        security_notes:
          type: array
          items:
            type: string
        metadata:
          type: object
...
  parameters:
...
    TaskIdParam:
      name: task_id
      in: path
      required: true
      schema:
        type: string
...
security:
  - ApiKeyAuth: []
paths:
...
  /think:
    post:
      tags: [Reasoning]
      summary: Generate solution
      description: Generate a solution for the given task.
      operationId: think
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskRequest"
      responses:
        "200":
          description: Solution generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SolutionResponse"
        "499":
          description: Task cancelled by user
        "503":
          description: Service starting
  /think/detail:
    post:
      tags: [Reasoning]
      summary: Full-detail solution
      description: Generate a full-detail solution.
      operationId: thinkDetail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskRequest"
      responses:
        "200":
          description: Full-detail solution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SolutionDetailResponse"
        "499":
          description: Task cancelled by user
        "503":
          description: Service starting
  /think/stream:
    post:
      tags: [Reasoning]
      summary: SSE streaming solution
      description: Generate a solution with Server-Sent Events (SSE) streaming.
      operationId: thinkStream
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskRequest"
      responses:
        "200":
          description: SSE event stream
          content:
            text/event-stream:
              schema:
                type: string
                description: SSE stream with step_start, step_end, complete events
        "503":
          description: Service starting
  /think/{task_id}:
    delete:
      tags: [Reasoning]
      summary: Cancel task
      description: Cancel a running think task by ID.
      operationId: cancelThink
      parameters:
        - $ref: "#/components/parameters/TaskIdParam"
      responses:
        "200":
          description: Cancellation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [cancelled, not_found_or_completed]
                  task_id:
                    type: string
...
  /knowledge/search:
    get:
      tags: [Knowledge]
      summary: Search knowledge
      description: Search knowledge base.
      operationId: knowledgeSearch
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: top_k
          in: query
          required: false
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object

  # ── Admin ──────────────────────────────────────────────────────
...